feat: 新增修改紀錄 table、白名單新增刪除都新增 log

+++ internal/repo/impl/BTMLoginToken.go
@@ -2,12 +2,18 @@ package impl
 
 import (
 	"BTM-backend/internal/repo/model"
+	"BTM-backend/pkg/error_code"
 
+	"github.com/go-kratos/kratos/v2/errors"
 	"gorm.io/gorm"
 	"gorm.io/gorm/clause"
 )
 
 func (repo *repository) IsLastLoginToken(db *gorm.DB, userID uint, loginToken string) (bool, error) {
+	if db == nil {
+		return false, errors.InternalServer(error_code.ErrDBError, "db is nil")
+	}
+
 	var count int64
 	if err := db.Model(&model.BTMLoginToken{}).
 		Where("user_id = ? AND login_token = ? AND deleted_at IS NULL", userID, loginToken).
@@ -18,6 +24,10 @@ func (repo *repository) IsLastLoginToken(db *gorm.DB, userID uint, loginToken st
 }
 
 func (repo *repository) CreateOrUpdateLastLoginToken(db *gorm.DB, userID uint, loginToken string) error {
+	if db == nil {
+		return errors.InternalServer(error_code.ErrDBError, "db is nil")
+	}
+
 	return db.Clauses(clause.OnConflict{
 		Columns: []clause.Column{{Name: "user_id"}},
 		DoUpdates: clause.Assignments(map[string]interface{}{
@@ -31,5 +41,9 @@ func (repo *repository) CreateOrUpdateLastLoginToken(db *gorm.DB, userID uint, l
 }
 
 func (repo *repository) DeleteLastLoginToken(db *gorm.DB, userID uint) error {
+	if db == nil {
+		return errors.InternalServer(error_code.ErrDBError, "db is nil")
+	}
+
 	return db.Delete(&model.BTMLoginToken{}, "user_id = ? ", userID).Error
 }
