+++ internal/repo/impl/whitelist.go
@@ -3,13 +3,19 @@ package impl
 import (
 	"BTM-backend/internal/domain"
 	"BTM-backend/internal/repo/model"
+	"BTM-backend/pkg/error_code"
 
+	"github.com/go-kratos/kratos/v2/errors"
 	"github.com/google/uuid"
 	"gorm.io/gorm"
 	"gorm.io/gorm/clause"
 )
 
 func (repo *repository) CreateWhitelist(db *gorm.DB, whitelist *domain.BTMWhitelist) error {
+	if db == nil {
+		return errors.InternalServer(error_code.ErrDBError, "db is nil")
+	}
+
 	modelWhitelist := WhitelistDomainToModel(*whitelist)
 	return db.Clauses(clause.OnConflict{
 		Columns: []clause.Column{{Name: "customer_id"}, {Name: "crypto_code"}, {Name: "address"}},
@@ -23,6 +29,10 @@ func (repo *repository) CreateWhitelist(db *gorm.DB, whitelist *domain.BTMWhitel
 }
 
 func (repo *repository) GetWhiteListByCustomerId(db *gorm.DB, customerID uuid.UUID, limit int, page int) ([]domain.BTMWhitelist, int64, error) {
+	if db == nil {
+		return nil, 0, errors.InternalServer(error_code.ErrDBError, "db is nil")
+	}
+
 	var modelWhitelists []model.BTMWhitelist
 	var total int64
 
@@ -44,15 +54,40 @@ func (repo *repository) GetWhiteListByCustomerId(db *gorm.DB, customerID uuid.UU
 }
 
 func (repo *repository) UpdateWhitelist(db *gorm.DB, whitelist *domain.BTMWhitelist) error {
+	if db == nil {
+		return errors.InternalServer(error_code.ErrDBError, "db is nil")
+	}
+
 	modelWhitelist := WhitelistDomainToModel(*whitelist)
 	return db.Save(&modelWhitelist).Error
 }
 
+func (repo *repository) GetWhiteListById(db *gorm.DB, id int64) (data domain.BTMWhitelist, err error) {
+	if db == nil {
+		return domain.BTMWhitelist{}, errors.InternalServer(error_code.ErrDBError, "db is nil")
+	}
+
+	var modelWhitelist model.BTMWhitelist
+	if err := db.First(&modelWhitelist, "id = ?", id).Error; err != nil {
+		return domain.BTMWhitelist{}, err
+	}
+
+	return WhitelistModelToDomain(modelWhitelist), nil
+}
+
 func (repo *repository) DeleteWhitelist(db *gorm.DB, id int64) error {
+	if db == nil {
+		return errors.InternalServer(error_code.ErrDBError, "db is nil")
+	}
+
 	return db.Delete(&model.BTMWhitelist{}, "id = ?", id).Error
 }
 
 func (repo *repository) SearchWhitelistByAddress(db *gorm.DB, customerID uuid.UUID, address string, limit int, page int) ([]domain.BTMWhitelist, int64, error) {
+	if db == nil {
+		return nil, 0, errors.InternalServer(error_code.ErrDBError, "db is nil")
+	}
+
 	var modelWhitelists []model.BTMWhitelist
 	var total int64
